--- src/amd.js
@@ -1,105 +1,0 @@
-const browserFakeUserAgent = "Fake/5.0 (FakeOS) AppleWebKit/123 (KHTML, like Gecko) Fake/12.3.4567.89 Fake/123.45";
-function amd(that) {
-    if (!that) {
-        return;
-    }
-    if (typeof window !== "undefined") {
-        return;
-    }
-    if (typeof global === "undefined") {
-        return;
-    }
-    global.navigator = {
-        userAgent: browserFakeUserAgent,
-        getUserMedia: function () { },
-    };
-    if (!global.console) {
-        global.console = {};
-    }
-    if (typeof global.console.debug === "undefined") {
-        global.console.debug =
-            global.console.info =
-                global.console.error =
-                    global.console.log =
-                        global.console.log ||
-                            function () {
-                                console.log(arguments);
-                            };
-    }
-    if (typeof document === "undefined") {
-        /*global document:true */
-        that.document = {};
-        document.createElement =
-            document.captureStream =
-                document.mozCaptureStream =
-                    function () {
-                        var obj = {
-                            getContext: function () {
-                                return obj;
-                            },
-                            play: function () { },
-                            pause: function () { },
-                            drawImage: function () { },
-                            toDataURL: function () {
-                                return "";
-                            },
-                        };
-                        return obj;
-                    };
-        document.addEventListener =
-            document.removeEventListener =
-                that.addEventListener =
-                    that.removeEventListener =
-                        function () { };
-        that.HTMLVideoElement = that.HTMLMediaElement = function () { };
-    }
-    if (typeof io === "undefined") {
-        that.io = function () {
-            return {
-                on: function (eventName, callback) {
-                    callback = callback || function () { };
-                    if (eventName === "connect") {
-                        callback();
-                    }
-                },
-                emit: function (eventName, data, callback) {
-                    callback = callback || function () { };
-                    if (eventName === "open-room" || eventName === "join-room") {
-                        callback(true, data.sessionid, null);
-                    }
-                },
-            };
-        };
-    }
-    if (typeof location === "undefined") {
-        /*global location:true */
-        that.location = {
-            protocol: "file:",
-            href: "",
-            hash: "",
-            origin: "self",
-        };
-    }
-    if (typeof screen === "undefined") {
-        /*global screen:true */
-        that.screen = {
-            width: 0,
-            height: 0,
-        };
-    }
-    if (typeof URL === "undefined") {
-        /*global screen:true */
-        that.URL = {
-            createObjectURL: function () {
-                return "";
-            },
-            revokeObjectURL: function () {
-                return "";
-            },
-        };
-    }
-    /*global window:true */
-    that.window = global;
-}
-export default amd(typeof global !== "undefined" ? global : null);
-//# sourceMappingURL=amd.js.map
