+++ tsconfig.json
@@ -5,7 +5,7 @@
     "module": "ESNext",
     "lib": ["ESNext", "DOM"],
     "moduleResolution": "Node",
-    "strict": false,
+    "strict": true,
     "sourceMap": true,
     "resolveJsonModule": true,
     "isolatedModules": false,
@@ -20,6 +20,5 @@
     "allowJs": true,
     "skipDefaultLibCheck": true
   },
-  "include": ["src"],
-  "file": ["global.d.ts"]
+  "include": ["src"]
 }
