+++ vite.config.docker.js
@@ -1,4 +1,4 @@
-import { defineConfig } from "vite";
+import defineConfig from "./vite.config";
 
 const path = require("path");
